Move makeClient to a shared module as well
Bug: https://phabricator.wikimedia.org/T140266
